Revert "Boilerplate wordpress cms"
