+++ data/posts/post-2.md
@@ -1,0 +1,14 @@
+---
+slug: 'slug-2'
+layout: 'post'
+title: 'Lorem ipsum dolor'
+date: '2021-03-01'
+---
+
+**Content goes here Ex. 2** 
+
+- Lorem ipsum _dolor sit amet_ consectetur adipiscing elit.
+- **Proin bibendum est justo**, molestie suscipit purus sagittis et. Nunc pulvinar accumsan fermentum. 
+
+Nunc convallis mi quis nisi rutrum luctus. Duis iaculis laoreet nunc, eu hendrerit tortor malesuada nec. Donec consectetur libero et sem condimentum, at fermentum velit ornare. Nullam ligula augue, rhoncus at molestie et, pulvinar sit amet sapien. Sed lorem neque, pulvinar nec vulputate eu, luctus id neque. Sed pulvinar, est lobortis tempor tincidunt, ligula ipsum commodo mi, in condimentum nisl dolor sed.
+
