+++ lib/posts.js
@@ -1,35 +1,63 @@
-import { INDEX_PAGE_QUERY, PAGES_QUERY, SINGLE_POST } from "./queries"
+import fs from 'fs'
+import path from 'path'
+import matter from 'gray-matter'
+import remark from 'remark'
+import html from 'remark-html'
 
-export async function getSinglePost(slug) {
-	return await fetch(process.env.WORDPRESS_GRAPHQL_ENDPOINT, {
-		method: 'POST',
-		headers: { 'Content-Type': 'application/json' },
-		body: JSON.stringify({
-			query: SINGLE_POST,
-			variables: {
-				id: slug,
-				idType: 'SLUG',
+
+const postsDirectory = path.join(process.cwd(), 'data/posts')
+
+export function getSortedPostsData() {
+	const fileNames = fs.readdirSync(postsDirectory)
+	const allPostsData = fileNames.map(fileName => {
+
+		const id = fileName.replace(/\.md$/, '')
+		const fullPath = path.join(postsDirectory, fileName)
+		const fileContents = fs.readFileSync(fullPath, 'utf8')
+		const parsedResult = matter(fileContents)
+
+		return {
+			id,
+			...parsedResult.data
+		}
+	})
+
+	return allPostsData.sort(({ date: a }, { date: b }) => {
+		if (a < b) {
+			return 1
+		} else if (a > b) {
+			return -1
+		} else {
+			return 0
+		}
+	})
+}
+
+export function getAllPostIds() {
+	const fileNames = fs.readdirSync(postsDirectory)
+	return fileNames.map(fileName => {
+		return {
+			params: {
+				id: fileName.replace(/\.md$/, '')
 			}
-		})
+		}
 	})
 }
 
-export async function getPostPages() {
-	return await fetch(process.env.WORDPRESS_GRAPHQL_ENDPOINT, {
-		method: 'POST',
-		headers: { 'Content-Type': 'application/json' },
-		body: JSON.stringify({
-			query: PAGES_QUERY
-		})
-	})
+export async function getPostData(id) {
+	const fullPath = path.join(postsDirectory, `${id}.md`)
+	const fileContents = fs.readFileSync(fullPath, 'utf8')
+
+	const parsedResult = matter(fileContents)
+	const processedContent = await remark()
+		.use(html)
+		.process(parsedResult.content)
+	const contentHtml = processedContent.toString()
+
+	return {
+		id,
+		contentHtml,
+		...parsedResult.data
+	}
 }
 
-export async function getAllPosts() {
-	return await fetch(process.env.WORDPRESS_GRAPHQL_ENDPOINT, {
-		method: 'POST',
-		headers: { 'Content-Type': 'application/json' },
-		body: JSON.stringify({
-			query: INDEX_PAGE_QUERY
-		})
-	})
-}
