+++ pages/index.js
@@ -1,18 +1,32 @@
 import Head from 'next/head'
 import Link from 'next/link'
 import Layout, { siteTitle } from '../components/layout'
-import { getAllPosts } from '../lib/posts'
 import utilStyles from '../styles/utils.module.css'
+import { getSortedPostsData } from '../lib/posts'
 
 export async function getStaticProps() {
-  const res = await getAllPosts()
-  const jsonResult = await res.json()
+  const allPostsData = getSortedPostsData()
   return {
     props: {
-      allPostsData: jsonResult.data.posts,
+      allPostsData,
     }
   }
 }
+
+/**
+ * export async function getServerSideProps(context) {
+    return {
+      props: {
+
+        Because getServerSideProps is called at request time, its parameter (context) contains request specific parameters.
+
+        You should use getServerSideProps only if you need to pre-render a page whose data must be fetched at request time.
+        Time to first byte (TTFB) will be slower than getStaticProps because the server must compute the result
+        on every request, and the result cannot be cached by a CDN without extra configuration.
+      }
+    }
+  }
+ */
 
 export default function Home(props) {
   return (
@@ -30,12 +44,10 @@ export default function Home(props) {
       <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
         <h2 className={utilStyles.headingLg}>Blog</h2>
         <ul className={utilStyles.list}>
-
-          {props.allPostsData.nodes.map(
-            (post) => (
-              <li className={utilStyles.listItem} key={post.slug}>
-                <Link href={"/posts/" + `${post.slug}`}>
-                  {post.title}
+          {props.allPostsData.map(({ id, title }) => (
+            <li className={utilStyles.listItem} key={id}>
+              <Link href={"/posts/" + `${id}`}>
+              {title}
               </Link>
             </li>
           ))}
