+++ pages/posts/[id].js
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 import Layout from '../../components/layout'
-import { getAllPostIds, getPostData } from '../../lib/locally-stored'
+import { getAllPostIds, getPostData } from '../../lib/posts'
 
 export async function getStaticProps({ params }) {
 	const postData = await getPostData(params.id)
@@ -28,10 +28,10 @@ export default function Post({ postData }) {
 
 	return <Layout>
 		<Head>
-			<title>Locally saved {postData.title}</title>
+			<title>{postData.title}</title>
 		</Head>
 		<br />
-		<h1>Locally saved post: {postData.id}</h1>
+		{postData.id}
 		<br />
 		{postData.date}
 		<div dangerouslySetInnerHTML={{ __html: postData.contentHtml }} />
