+++ public/admin/config.yml
@@ -1,0 +1,51 @@
+backend:
+  name: git-gateway
+  branch: master
+media_folder: public/assets
+public_folder: /assets
+publish_mode: editorial_workflow
+
+collections:
+  - name: "post"
+    label: "Post"
+    folder: "data/posts/"
+    extension: "md"
+    format: "frontmatter"
+    create: true
+    slug: "{{slug}}"
+    summary: "{{title}}"
+    fields:
+      - label: "Slug"
+        name: "slug"
+        widget: "string"
+      - label: "Layout"
+        name: "layout"
+        widget: "string"
+      - label: "Title"
+        name: "title"
+        widget: "string"
+      - label: "Publish Date"
+        name: "date"
+        widget: "datetime"
+        format: "YYYY-MM-DD"
+        dateFormat: "YYYY-MM-DD"
+        timeFormat: false
+      # - label: Tags
+      #   label_singular: "Tag"
+      #   name: "tags"
+      #   widget: list
+      #   summary: "{{fields.tag}}"
+      #   field:
+      #     label: Tag
+      #     name: tag
+      #     widget: relation
+      #     collection: "meta"
+      #     file: "tags"
+      #     searchFields:
+      #       - "tags.*.name"
+      #     displayFields:
+      #       - "tags.*.name"
+      #     valueField: "tags.*.slug"
+      - label: "Body"
+        name: "body"
+        widget: "markdown"
